when the call his ended this time, The app returns to Main Activity, I think that works for now, After todays meeting, I'll see if returning to SecondActivity.kt can be implemented.
I made use of chatGPT, for reasearch, so it contains deprecated codes
